centralize util functions
