+++ core-web/libs/sdk/react/src/lib/next/components/Column/Column.tsx
@@ -1,7 +1,8 @@
+import { combineClasses, getColumnPositionClasses } from '@dotcms/uve/internal';
+import { DotPageAssetLayoutColumn } from '@dotcms/uve/types';
+
 import styles from './Column.module.css';
 
-import { DotPageAssetLayoutColumn } from '../../types';
-import { combineClasses, getColumnPositionClasses } from '../../utils';
 import { Container } from '../Container/Container';
 
 /**
