+++ core-web/libs/sdk/react/src/lib/next/components/Row/Row.tsx
@@ -1,6 +1,7 @@
+import { DotPageAssetLayoutRow } from '@dotcms/uve/types';
+
 import styles from './Row.module.css';
 
-import { DotPageAssetLayoutRow } from '../../types';
 import { Column } from '../Column/Column';
 
 /**
