+++ core-web/libs/sdk/uve/src/types.ts
@@ -1,2 +1,3 @@
 export * from './lib/types/editor/public';
 export * from './lib/types/events/public';
+export * from './lib/types/page/public';
